requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
langchain-core>=0.1.0
langsmith>=0.0.70
faiss-cpu>=1.7.0
chromadb>=0.4.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
streamlit>=1.28.0
requests>=2.31.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
docx2txt>=0.8 